v 9.0
- Незначні зміни інтерфейсу
- Додана можливість обчислення рівняння методом хорд
- Додана можливість побудови графіка функції
- Виправлення багів

+++ EquationСalculationCoursework/Data.cs
@@ -12,8 +12,8 @@ namespace EquationСalculationCoursework
 {
     class Data
     {
-        public static double a { get; set; }
-        public static double b { get; set; }
+        public static int a { get; set; }
+        public static int b { get; set; }
         public static double c { get; set; }
         public static double eps { get; set; }
         public static double A { get; set; }
